ex 8 started
