+++ src/main/java/com/luxoft/bankapp/commands/GetAccountsCommand.java
