+++ src/main/java/com/luxoft/bankapp/exceptions/FeedException.java
